+++ src/plants/history_plant.rs
@@ -576,9 +576,9 @@ impl RithmicHistoryPlantHandle {
             .next()
             .ok_or(RithmicError::EmptyResponse)?;
 
-        if let Some(err) = response.error {
+        if let Some(err) = response.request_error() {
             error!("history_plant: login failed {:?}", err);
-            Err(RithmicError::ServerError(err))
+            Err(err)
         } else {
             let _ = self.sender.send(HistoryPlantCommand::SetLogin).await;