+++ src/plants/order_plant.rs
@@ -1244,9 +1244,9 @@ impl RithmicOrderPlantHandle {
             .next()
             .ok_or(RithmicError::EmptyResponse)?;
 
-        if let Some(err) = response.error {
+        if let Some(err) = response.request_error() {
             error!("order_plant: login failed {:?}", err);
-            Err(RithmicError::ServerError(err))
+            Err(err)
         } else {
             let _ = self.sender.send(OrderPlantCommand::SetLogin).await;