+++ src/plants/pnl_plant.rs
@@ -445,9 +445,9 @@ impl RithmicPnlPlantHandle {
             .next()
             .ok_or(RithmicError::EmptyResponse)?;
 
-        if let Some(err) = response.error {
+        if let Some(err) = response.request_error() {
             error!("pnl_plant: login failed {:?}", err);
-            Err(RithmicError::ServerError(err))
+            Err(err)
         } else {
             let _ = self.sender.send(PnlPlantCommand::SetLogin).await;